requirements/requirements.txt
# SHA1:35155ae60e9de348b3cef0374eea410b14b2e3dd
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via starlette
attrs==24.2.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.12.3
    # via -r requirements\requirements.in
blinker==1.8.2
    # via selenium-wire
brotli==1.1.0
    # via selenium-wire
certifi==2024.8.30
    # via
    #   requests
    #   selenium
    #   selenium-wire
cffi==1.17.1
    # via
    #   cryptography
    #   trio
charset-normalizer==3.4.0
    # via requests
chromedriver==2.24.1
    # via -r requirements\requirements.in
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==43.0.3
    # via pyopenssl
fastapi==0.115.4
    # via -r requirements\requirements.in
h11==0.14.0
    # via
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via selenium-wire
hpack==4.0.0
    # via h2
hyperframe==6.0.1
    # via
    #   h2
    #   selenium-wire
idna==3.10
    # via
    #   anyio
    #   requests
    #   trio
kaitaistruct==0.10
    # via selenium-wire
outcome==1.3.0.post0
    # via trio
packaging==24.1
    # via webdriver-manager
pillow==11.0.0
    # via -r requirements\requirements.in
pyasn1==0.6.1
    # via selenium-wire
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pydivert==2.1.0
    # via selenium-wire
pyopenssl==24.2.1
    # via selenium-wire
pyparsing==3.2.0
    # via selenium-wire
pysocks==1.7.1
    # via
    #   selenium-wire
    #   urllib3
python-dotenv==1.0.1
    # via webdriver-manager
requests==2.32.3
    # via
    #   -r requirements\requirements.in
    #   webdriver-manager
selenium==4.26.1
    # via
    #   -r requirements\requirements.in
    #   selenium-wire
selenium-wire==5.1.0
    # via -r requirements\requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.6
    # via beautifulsoup4
starlette==0.41.2
    # via fastapi
tqdm==4.66.6
    # via -r requirements\requirements.in
trio==0.27.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   selenium
urllib3[socks]==2.2.3
    # via
    #   requests
    #   selenium
uvicorn==0.32.0
    # via -r requirements\requirements.in
webdriver-manager==4.0.2
    # via -r requirements\requirements.in
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via
    #   selenium-wire
    #   trio-websocket
yt-dlp==2024.11.4
    # via -r requirements\requirements.in
zstandard==0.23.0
    # via selenium-wire
